Folder structure change to conform to the blueprint
