+++ app.py
@@ -1,0 +1,23 @@
+
+from app import create_app
+
+
+# version1 = Blueprint("apiv1", __name__, url_prefix="/api/v1")
+
+app= create_app()
+
+
+
+
+
+# MY DECORATORS
+
+    # decorator where a user requires a token to view
+
+
+
+
+    # a registered user can ask a question
+
+if __name__ == '__main__':
+        app.run(debug=True)
