+++ app/API/v1/models/model.py
@@ -3,6 +3,7 @@ from flask import Flask, request, jsonify
 
 users= []
 questions= []
+answers=[]
